deep_research/requirements.txt
# deep_research - 딥 리서치 (LangGraph + Tavily)
langchain-openai>=0.2.0
langchain-tavily>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
